+++ pgvector/asyncpg/__init__.py
@@ -1,4 +1,4 @@
-from ..utils import from_db, from_db_binary, to_db, to_db_binary
+from ..utils import from_db, from_db_binary, to_db, to_db_binary, HalfVec, SparseVec
 
 __all__ = ['register_vector']
 
@@ -8,5 +8,20 @@ async def register_vector(conn):
         'vector',
         encoder=to_db_binary,
         decoder=from_db_binary,
+        format='binary'
+    )
+
+    await conn.set_type_codec(
+        'halfvec',
+        encoder=lambda v: HalfVec(v).to_db_binary(),
+        decoder=HalfVec.from_db_binary,
+        format='binary'
+    )
+
+    await conn.set_type_codec(
+        'sparsevec',
+        # TODO fix
+        encoder=lambda v: isinstance(v, SparseVec) and v.to_db_binary(),
+        decoder=SparseVec.from_db_binary,
         format='binary'
     )